uni_image/requirement.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.3
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-redis==4.9.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
future==0.16.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
pbr==4.0.4
Pillow==5.3.0
pycryptodome==3.6.3
PyJWT==1.6.4
pytz==2018.4
redis==2.10.6
requests==2.19.1
six==1.11.0
stevedore==1.28.0
uritemplate==3.0.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
